Added a state machine for the robot
The state machine can be used to write the behavior of the robot in a simple, safe and flexible way.
In the FSM is currently only a single state: State_Idle. This state returns (by means of example) a new instance of the same state when update is called and logs all the state transtions.

! This first commit contains a working WiP framework with lots of dummy values.
! Code still needs to implement the ROS parts and needs to be splitted into header files.